+++ api/.irb_history
@@ -57,3 +57,9 @@ Like.all
 \%\
 exit
 exit
+Like.all
+Like.all
+Like.all
+Link.all
+Like.all
+exit
